Added comments from @Noggling
